Renaming keyword.ts to search-query-normalizer.ts
