+++ medicines/web/src/services/search-query-normalizer.ts
